+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## [dev]
 - add tcRank, tcRep, fncV3Rep and fncV3Rank to daily_model_performances
+- add icRep and icRank to SignalsAPI.daily_model_performances
 - add tc and tcPercentile to daily_submissions_performances
 - make `stake_set` work with multi model accounts
 - add tc to the round_model_performances method (PR #74)